pyproject.toml
[tool.poetry]
name = "task-manager-zhukata-bot"
version = "0.1.0"
description = ""
authors = ["zhukata <zhukata@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.13.1"
python-dotenv = "^1.0.1"
asyncio = "^3.4.3"
sqlalchemy = "^2.0.36"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
